Change sample_no to a uint64_t
The sample_no input to the zap_accumulate function was an int
that would overflow after 2B samples. This changes the type to uint64_t.

+++ lib/src/zap/zap_priv.h
@@ -329,7 +329,7 @@ int zap_event_add(struct zap_event_queue *q, zap_ep_t ep, void *ctxt);
  */
 struct zap_thrstat {
 	char *name;
-	int window_size;
+	uint64_t window_size;
 	struct timespec wait_start;
 	struct timespec wait_end;
 	int waiting;
